+++ RELEASE-NOTES.md
@@ -2,7 +2,7 @@
 Release Notes
 =============
 
-### Upcoming Changes
+### Jumi 0.5.437 (2014-02-07)
 
 - Workaround for Eclipse's JUnit integration producing non-standard classpath elements on Windows ([Issue #6](https://github.com/orfjackal/jumi/issues/6))
 - Deprecated the `SuiteConfiguration.classPath` property. Use `classpath` instead. The old methods will be removed after about 180 days